+++ components/shared/rightSocial.tsx
@@ -15,7 +15,7 @@ const RightSocialComponent = () => {
             <div className='relative hover:text-gray-950 cursor-pointer transition-all ease-linear'>
                 <a href="mailto:shahbaazshaik78@gmail.com" className='absolute text-gray-600 -top-32 -left-[6.8rem] rotate-90 text-base '>shahbaazshaik78@gmail.com</a>
             </div>
-            <span className='bg-gray-950 h-40 w-[2px]  mt-8 rounded-full hidden sm:block'></span>
+            <span className='bg-gray-500 h-40 w-[2px]  mt-8 rounded-full hidden sm:block'></span>
         </motion.div>
     )
 }