+++ changelog.md
@@ -1,3 +1,13 @@
+CL-4.0.6: Added Header Animation, Refactoring & Code Optimizations.
+- Decreased color contrast of left/right social components.
+- Added ids to components so that we can scroll down to particular section upon clicking header.
+- Updated Hero section intro.
+- Updated scroll behaviour to smooth.
+- Added scroll-margin-top to make sure upon clicking menu links every section's content is properly aligned and visible. 
+- Added custom hooks (useSectionInView) to track the ref/inView status and (useActiveSectionContext) to handle context provider exceptions.
+- installed 3rd party libraries like **react-intersection-observer** to track the html element is in view or not. And Installed **clsx** to add conditional styling to header animations. 
+- Added interfaces and types wherever required.
+
 CL-4.0.5: Added Projects.
 - Added Project component.
 - Added Project assets and types.